+++ novawallet/Modules/Vote/Governance/SwipeGov/SwipeGov/Views/CardsZStack/CardStackable.swift
@@ -1,0 +1,6 @@
+protocol CardStackable: AnyObject {
+    func didBecomeTopView()
+    func didAddToStack()
+    func didPopFromStack(direction: CardsZStack.DismissalDirection)
+    func prepareForReuse()
+}
