+++ novawallet/Modules/Vote/Governance/SwipeGov/SwipeGov/Views/CardsZStack/CardsZStack.swift
@@ -1,44 +1,13 @@
 import UIKit
 import SnapKit
 
-protocol CardStackable: AnyObject {
-    func didBecomeTopView()
-    func didAddToStack()
-    func didPopFromStack(direction: CardsZStack.DismissalDirection)
-    func prepareForReuse()
-}
-
-struct CardsZStackViewModel {
-    let allCards: [VoteCardId: VoteCardViewModel]
-    let changeModel: CardsZStackChangeModel
-    let emptyViewModel: SwipeGovEmptyStateViewModel
-    let validationAction: ((VoteCardViewModel?) -> Bool)?
-
-    var stackIsEmpty: Bool {
-        allCards.isEmpty
-    }
-}
-
-struct CardsZStackChangeModel {
-    let inserts: [VoteCardViewModel]
-    let updates: [VoteCardId: VoteCardViewModel]
-    let deletes: Set<VoteCardId>
-}
-
 final class CardsZStack: UIView {
-    enum DismissalDirection {
-        case left
-        case right
-        case top
-        case bottom
-    }
-
     private let maxCardsAlive: Int
-    private var stackedViews: [VoteCardView] = []
-    private var dismissingIds = Set<VoteCardId>()
-    private var emptyStateView: UIView?
-    private var viewPool: [VoteCardView] = []
-    private var viewModelsQueue: [VoteCardViewModel] = []
+    private(set) var stackedViews: [VoteCardView] = []
+    private(set) var dismissingIds = Set<VoteCardId>()
+    private(set) var emptyStateView: UIView?
+    private(set) var viewPool: [VoteCardView] = []
+    private(set) var viewModelsQueue: [VoteCardViewModel] = []
 
     private var validateAction: ((VoteCardViewModel?) -> Bool)?
 
@@ -176,127 +145,6 @@ final class CardsZStack: UIView {
         }
 
         manageStack()
-    }
-}
-
-// MARK: - Animate
-
-private extension CardsZStack {
-    func animateStackManage() {
-        stackedViews.reversed().enumerated().forEach { index, view in
-            let (scale, scaleTransform): (CGFloat, CGAffineTransform) = createScaleTransform(
-                for: index
-            )
-            let translationTransform = createTranslationTransform(
-                cardIndex: index,
-                cardScale: scale,
-                insertLast: false
-            )
-            let finalTransform = scaleTransform.concatenating(translationTransform)
-
-            UIView.animate(
-                withDuration: Constants.StackManagingAnimation.duration,
-                delay: Constants.StackManagingAnimation.delay,
-                usingSpringWithDamping: Constants.StackManagingAnimation.springDamping,
-                initialSpringVelocity: Constants.StackManagingAnimation.springVelocity,
-                options: [.curveEaseInOut]
-            ) {
-                view.transform = finalTransform
-            }
-        }
-    }
-
-    func animateCardAdd(_ cardView: UIView) {
-        let (scale, scaleTransform): (CGFloat, CGAffineTransform) = createScaleTransform(
-            for: stackedViews.endIndex
-        )
-        let translationTransform = createTranslationTransform(
-            cardIndex: stackedViews.endIndex,
-            cardScale: scale,
-            insertLast: true
-        )
-
-        cardView.alpha = 0
-        cardView.transform = scaleTransform
-
-        UIView.animate(
-            withDuration: Constants.CardAddAnimation.duration,
-            delay: Constants.CardAddAnimation.delay,
-            usingSpringWithDamping: Constants.CardAddAnimation.springDamping,
-            initialSpringVelocity: Constants.CardAddAnimation.springVelocity,
-            options: [.curveEaseInOut]
-        ) {
-            cardView.alpha = 1
-            cardView.transform = scaleTransform.concatenating(translationTransform)
-        }
-    }
-
-    func animateCardDismiss(
-        _ topCard: UIView,
-        direction: DismissalDirection,
-        completion: @escaping () -> Void
-    ) {
-        let translation: (x: CGFloat, y: CGFloat) = switch direction {
-        case .left: (-1.5 * bounds.width, 0)
-        case .right: (1.5 * bounds.width, 0)
-        case .top: (0, -1.5 * bounds.height)
-        case .bottom: (0, 1.5 * bounds.height)
-        }
-
-        let rotationDirection: CGFloat = switch direction {
-        case .left: -1
-        case .right: 1
-        case .top: 0
-        case .bottom: 0
-        }
-
-        UIView.animate(
-            withDuration: Constants.CardDismissAnimation.duration,
-            animations: {
-                topCard.transform = CGAffineTransform(rotationAngle: 0.15 * rotationDirection)
-                    .translatedBy(x: translation.x, y: translation.y)
-            },
-            completion: { _ in
-                completion()
-            }
-        )
-    }
-
-    func createScaleTransform(
-        for cardIndex: Int
-    ) -> (scale: CGFloat, transform: CGAffineTransform) {
-        let scale = NSDecimalNumber(
-            decimal: pow(Decimal(Constants.stackZScaling), cardIndex)
-        ).doubleValue
-
-        let transform = CGAffineTransform(
-            scaleX: scale,
-            y: scale
-        )
-
-        return (scale, transform)
-    }
-
-    func createTranslationTransform(
-        cardIndex: Int,
-        cardScale: CGFloat,
-        insertLast: Bool
-    ) -> CGAffineTransform {
-        guard let baseHeight = stackedViews.last?.frame.height else {
-            return .identity
-        }
-
-        let prevIndex = cardIndex > 0
-            ? cardIndex - 1
-            : cardIndex
-
-        let heightDelta = (baseHeight - (baseHeight * cardScale)) / 2
-        let translationY = heightDelta + Constants.lowerCardsOffset * CGFloat(insertLast ? prevIndex : cardIndex)
-
-        return CGAffineTransform(
-            translationX: .zero,
-            y: translationY
-        )
     }
 }
 
@@ -479,47 +327,16 @@ private extension CardsZStack {
     }
 }
 
-private extension CardsZStack {
-    enum Constants {
-        enum CardIdentityAnimation {
-            static let duration: CGFloat = 0.35
-            static let delay: CGFloat = .zero
-            static let springDamping: CGFloat = 0.75
-            static let springVelocity: CGFloat = 0.6
-        }
+extension CardsZStack {
+    enum DismissalDirection {
+        case left
+        case right
+        case top
+        case bottom
+    }
 
-        enum CardDismissAnimation {
-            static let duration: CGFloat = 0.35
-        }
-
-        enum CardAddAnimation {
-            static let duration: CGFloat = 0.35
-            static let delay: CGFloat = .zero
-            static let springDamping: CGFloat = 0.75
-            static let springVelocity: CGFloat = 0.6
-        }
-
-        enum StackManagingAnimation {
-            static let duration: CGFloat = 0.3
-            static let delay: CGFloat = .zero
-            static let springDamping: CGFloat = 0.6
-            static let springVelocity: CGFloat = 0.3
-        }
-
-        enum EmptyStateAnimation {
-            static let startDuration: CGFloat = 0.2
-            static let endDuration: CGFloat = 0.1
-        }
-
-        static let emptyStateHideScaling: CGFloat = 0.8
-        static let emptyStateShowScaling: CGFloat = 1.05
-        static let cardCornerRadius: CGFloat = 16
-        static let stackZScaling: CGFloat = 0.9
-        static let lowerCardsOffset: CGFloat = 10.0
-        static let horizontalScreenSizeDivider: CGFloat = 3.15
-        static let verticalScreenSizeDivider: CGFloat = 3.8
-        static let topMostY = -(UIScreen.main.bounds.height / Constants.verticalScreenSizeDivider)
-        static let leftMostX = -(UIScreen.main.bounds.width / Constants.horizontalScreenSizeDivider)
-        static let rightMostX = UIScreen.main.bounds.width / Constants.horizontalScreenSizeDivider
+    struct Actions {
+        let emptyViewAction: () -> Void
+        let validationClosure: (VoteCardViewModel?) -> Bool
     }
 }
