+++ novawallet/Modules/Vote/Governance/SwipeGov/SwipeGov/Views/CardsZStack/CardsZStackViewModel.swift
@@ -1,0 +1,16 @@
+struct CardsZStackViewModel {
+    let allCards: [VoteCardId: VoteCardViewModel]
+    let changeModel: CardsZStackChangeModel
+    let emptyViewModel: SwipeGovEmptyStateViewModel
+    let validationAction: ((VoteCardViewModel?) -> Bool)?
+
+    var stackIsEmpty: Bool {
+        allCards.isEmpty
+    }
+}
+
+struct CardsZStackChangeModel {
+    let inserts: [VoteCardViewModel]
+    let updates: [VoteCardId: VoteCardViewModel]
+    let deletes: Set<VoteCardId>
+}
