+++ novawallet/Modules/Vote/Governance/SwipeGov/SwipeGov/Views/CardsZStack/VoteCard/StackCardUpdatable.swift
