+++ novawallet/Modules/Vote/Governance/SwipeGov/SwipeGovVotingList/SwipeGovVotingListProtocols.swift
@@ -27,7 +27,10 @@ protocol SwipeGovVotingListInteractorOutputProtocol: AnyObject {
     func didReceive(_ error: SwipeGovVotingListInteractorError)
 }
 
-protocol SwipeGovVotingListWireframeProtocol: AlertPresentable, ErrorPresentable, CommonRetryable, SwipeGovAlertPresentable {
+protocol SwipeGovVotingListWireframeProtocol: AlertPresentable, 
+                                                ErrorPresentable,
+                                              CommonRetryable,
+                                                SwipeGovAlertPresentable {
     func close(view: ControllerBackedProtocol?)
     func showSetup(
         from view: ControllerBackedProtocol?,
